requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
fastapi[all]==0.104.1
uvicorn==0.24.0